+++ test/strojure/zizzmap/impl_test.clj
@@ -12,7 +12,7 @@
 
 ;;,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 
-(def ^:private -e (impl/boxed-map-entry :k (impl/boxed-value (identity :x))))
+(def ^:private -e (impl/boxed-map-entry :k (impl/boxed-delay :x)))
 
 (deftest boxed-map-entry-t
   (test/are [expr result] (= result expr)
@@ -41,7 +41,7 @@
 
     (assoc -e 0 :kk) #_= [:kk :x]
     (assoc -e 1 :xx) #_= [:k :xx]
-    (assoc -e 1 (impl/boxed-value (identity :xx))) #_= [:k :xx]
+    (assoc -e 1 (impl/boxed-delay :xx)) #_= [:k :xx]
     (assoc -e 2 :y) #_= [:k :x :y]
 
     (contains? -e 0) #_= true
@@ -52,14 +52,14 @@
     (.entryAt ^IPersistentVector -e 1) #_= [1 :x]
     (.entryAt ^IPersistentVector -e 2) #_= nil
     (let [a (atom :pending)
-          e (impl/boxed-map-entry :k (impl/boxed-value (do (reset! a :realized)
-                                                           :x)))
+          e (impl/boxed-map-entry :k (impl/boxed-delay (reset! a :realized)
+                                                       :x))
           e (.entryAt ^IPersistentVector e 1)]
       [e @a]) #_= [[1 :x] :pending]
 
     (.assocN ^IPersistentVector -e 0 :kk) #_= [:kk :x]
     (.assocN ^IPersistentVector -e 1 :xx) #_= [:k :xx]
-    (.assocN ^IPersistentVector -e 1 (impl/boxed-value (identity :xx))) #_= [:k :xx]
+    (.assocN ^IPersistentVector -e 1 (impl/boxed-delay :xx)) #_= [:k :xx]
     (.assocN ^IPersistentVector -e 2 :y) #_= [:k :x :y]
 
     (empty -e) #_= (empty (MapEntry. :k :x))
@@ -69,18 +69,18 @@
     (second (seq -e)) #_= :x
 
     (let [a (atom :pending)
-          e (impl/boxed-map-entry :k (impl/boxed-value (do (reset! a :realized)
-                                                           :x)))]
+          e (impl/boxed-map-entry :k (impl/boxed-delay (reset! a :realized)
+                                                       :x))]
       [(first e) @a]) #_= [:k :pending]
 
     (let [a (atom :pending)
-          e (impl/boxed-map-entry :k (impl/boxed-value (do (reset! a :realized)
-                                                           :x)))]
+          e (impl/boxed-map-entry :k (impl/boxed-delay (reset! a :realized)
+                                                       :x))]
       [(second e) @a]) #_= [:x :realized]
 
     (.equiv ^IPersistentVector -e -e) #_= true
     (.equiv ^IPersistentVector -e [:k :x]) #_= true
-    (.equiv ^IPersistentVector -e (impl/boxed-map-entry :k (impl/boxed-value (identity :x)))) #_= true
+    (.equiv ^IPersistentVector -e (impl/boxed-map-entry :k (impl/boxed-delay :x))) #_= true
     (.equiv ^IPersistentVector -e [:k :y]) #_= false
 
     (sequential? -e) #_= true
@@ -102,7 +102,7 @@
 ;;,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 
 (def ^:private -m
-  (impl/persistent-map {:a (impl/boxed-value (identity :x))
+  (impl/persistent-map {:a (impl/boxed-delay :x)
                         :b :y}))
 
 (deftest persistent-map-t
@@ -129,8 +129,8 @@
     (= -m -m) #_= true
     (= -m {:a :x :b :y}) #_= true
     (= {:a :x :b :y} -m) #_= true
-    (= (impl/persistent-map {:a (impl/boxed-value (identity :x))})
-       (impl/persistent-map {:a (impl/boxed-value (identity :x))})) #_= true
+    (= (impl/persistent-map {:a (impl/boxed-delay :x)})
+       (impl/persistent-map {:a (impl/boxed-delay :x)})) #_= true
 
     (assoc -m :a :xx) #_= {:a :xx :b :y}
     (assoc -m :b :yy) #_= {:a :x :b :yy}
@@ -170,7 +170,7 @@
     (set (vals -m)) #_= #{:x :y}
 
     (conj -m [:c :z]) #_= {:a :x, :b :y, :c :z}
-    (conj -m [:c (impl/boxed-value (identity :z))]) #_= {:a :x, :b :y, :c :z}
+    (conj -m [:c (impl/boxed-delay :z)]) #_= {:a :x, :b :y, :c :z}
 
     (empty -m) #_= {}
     (impl/persistent? (empty -m)) #_= true
@@ -186,106 +186,106 @@
   (testing "Value laziness"
     (test/are [expr result] (= result expr)
 
-      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-value
-                                                           (do (reset! a :realized)
-                                                               :x))})
+      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-delay
+                                                           (reset! a :realized)
+                                                           :x)})
             before @a
             x (get m :a)
             after @a]
         [before x after]) #_= [:pending :x :realized]
 
-      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-value
-                                                           (do (reset! a :realized)
-                                                               :x))})
+      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-delay
+                                                           (reset! a :realized)
+                                                           :x)})
             m (assoc m :a :xx)]
         [@a m]) #_= [:pending {:a :xx}]
 
-      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-value
-                                                           (do (reset! a :realized)
-                                                               :x))})
+      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-delay
+                                                           (reset! a :realized)
+                                                           :x)})
             m (assoc m :b :yy)]
         [@a m]) #_= [:pending {:a :x :b :yy}]
 
-      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-value
-                                                           (do (reset! a :realized)
-                                                               :x))
+      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-delay
+                                                           (reset! a :realized)
+                                                           :x)
                                                       :b :y})
             m (dissoc m :a)]
         [@a m]) #_= [:pending {:b :y}]
 
-      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-value
-                                                           (do (reset! a :realized)
-                                                               :x))
+      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-delay
+                                                           (reset! a :realized)
+                                                           :x)
                                                       :b :y})
             m (dissoc m :b)]
         [@a m]) #_= [:pending {:a :x}]
 
-      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-value
-                                                           (do (reset! a :realized)
-                                                               :x))
+      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-delay
+                                                           (reset! a :realized)
+                                                           :x)
                                                       :b :y})
             m (update m :b name)]
         [@a m]) #_= [:pending {:a :x :b "y"}]
 
-      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-value
-                                                           (do (reset! a :realized)
-                                                               :x))
+      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-delay
+                                                           (reset! a :realized)
+                                                           :x)
                                                       :b :y})
             m (select-keys m [:a :b])]
         [@a m]) #_= [:realized {:a :x :b :y}]
 
-      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-value
-                                                           (do (reset! a :realized)
-                                                               :x))
+      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-delay
+                                                           (reset! a :realized)
+                                                           :x)
                                                       :b :y})
             m (select-keys m [:b])]
         [@a m]) #_= [:pending {:b :y}]
 
-      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-value
-                                                           (do (reset! a :realized)
-                                                               :x))
+      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-delay
+                                                           (reset! a :realized)
+                                                           :x)
                                                       :b :y})
             _ (doall (seq m))]
         [@a]) #_= [:pending]
 
-      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-value
-                                                           (do (reset! a :realized)
-                                                               :x))
+      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-delay
+                                                           (reset! a :realized)
+                                                           :x)
                                                       :b :y})
             m (into m {:c :z})]
         [@a m]) #_= [:pending {:a :x, :b :y, :c :z}]
 
-      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-value
-                                                           (do (reset! a :realized)
-                                                               :x))
+      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-delay
+                                                           (reset! a :realized)
+                                                           :x)
                                                       :b :y})
             m (into {:c :z} m)]
         [@a m]) #_= [:realized {:a :x, :b :y, :c :z}]
 
-      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-value
-                                                           (do (reset! a :realized)
-                                                               :x))
+      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-delay
+                                                           (reset! a :realized)
+                                                           :x)
                                                       :b :y})
             m (conj m [:c :z])]
         [@a m]) #_= [:pending {:a :x, :b :y, :c :z}]
 
-      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-value
-                                                           (do (reset! a :realized)
-                                                               :x))
+      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-delay
+                                                           (reset! a :realized)
+                                                           :x)
                                                       :b :y})
             k (key (find m :a))]
         [@a k]) #_= [:pending :a]
 
-      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-value
-                                                           (do (reset! a :realized)
-                                                               :x))
+      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-delay
+                                                           (reset! a :realized)
+                                                           :x)
                                                       :b :y})
             v (val (find m :a))]
         [@a v]) #_= [:realized :x]
 
-      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-value
-                                                           (do (reset! a :realized)
-                                                               :x))
+      (let [a (atom :pending) m (impl/persistent-map {:a (impl/boxed-delay
+                                                           (reset! a :realized)
+                                                           :x)
                                                       :b :y})
             m (persistent! (transient m))]
         [@a m]) #_= [:pending {:a :x, :b :y}]